+++ CHANGELOG.md
@@ -6,9 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Notification when timer reached x seconds from game start
 ### Fixed
 - Cell table highlighted when entering game
-- Guardians not highlighted with point balance when level too high
+- Guardians not highlighted with point balance helper when level too high
 
 
 ## [1.2.0]